+++ changelog.md
@@ -1,3 +1,8 @@
+2.2.0
+==========
+- Re-add netcoreapp3.1 target framework.
+- Update to Microsoft.Identity.Abstractions 2.0.1 which includes API change moving from `IDownstreamRestApi` to `IDownstreamApi`.
+
 2.1.0
 ==========
 [Id Web v2.1.0](https://github.com/AzureAD/microsoft-identity-web/wiki/v2.0) brings a varity of new higher-level APIs, including support for .NET Framework (Owin), Daemon scenarios, and the new DownstreamRestApi.